Automated white-space checks and fixes.
As of now, the script is never run automatically but must be invoked
explicitly.  You can run

    $ python3 extras/checks/whitespace.py --help

for more information about how it can be used.  All checks are
conservative and aim for zero false positives (at the cost of many
false negatives).  What this means is: if the script reports an issue,
go ahead and fix it.

	* .gitignore: Sorted and formatted.
	* CMakeLists.txt: Fixed white-space issues.
	* README.md: Likewise.
	* doxygen.in: Likewise.
	* extras/checks/whitespace.py: New.