Also check vertical white-space.
I thought this would be an easy exercise but it turned out I had to do some
major refactoring.  We're now reading the whole file into memory upfront (which
should be okay in this case).  The code was also cleaned up during the
refactoring and should now be easier to extend.

	* extras/checks/whitespace.py: Enhance and refactor.