Complete the full list of token types.
At the moment, this is very repetitive.  Some of this redundancy could be
avoided by using techniques like X-macros but it is questionable whether this
would make the code cleaner and Doxygen wouldn't be able to generate
documentation for it.  The issue is also mitigated by the compiler being able
to do fairly good consistency checks accross the duplications.

	* src/lexer/keyword.cpp:
	(classify_word): Implement.
	* src/lexer/token.cpp: Tweak stream-insertion operator.
	* src/lexer/token.hpp: Doxygen.
	* src/lexer/token_iterator.hpp: Doxygen.
	* src/lexer/token_type.cpp: Implement.
	* src/lexer/token_type.hpp:
	(all_token_types): New.
	(category): New.
	(token_category): New.
	(token_type): Add all types.
	* src/lexer/token_type.tpp: Implement.
	* tests/lexer/keyword.cpp: Better tests.
	* tests/lexer/token.cpp: Test stream-insertion operator.