A very basic unit test for the lexer.
This test checks that each punctuation token is lexed correctly in isolation.
It can also check the same for isolated keywords but since keywords are not
implemented in the lexer yet, this is currently commented-out from the test.
Un-comment the obvious to also check keywords.  The test uses a
`std::forward_list` as input.  Invoking undefined behavior by dereferencing a
past-the-end iterator for a `std::string` often goes by unnoticed because
you're lucky to hit the NUL byte.  It is undefined behavior nonetheless.  Our
actual compiler will invoke the lexer with input iterators which are even
weaker then the forward iterators provided by `std::forward_list`.

	* tests/lexer/lexer.cpp: Added one unit test.