Update the CLI to support --lextest as well as --echo.
This changes the semantic model of the CLI.  The general usage pattern is now

    $ minijava [STAGE] [--output=OUT] [IN]

where *STAGE* is an option that identifies the compilation stage at which to
intercept, such as `--echo` or `--lextest`.  *OUT* and *IN* are file names that
-- if not provided -- default to standard output and standard input
respectively.  The special file name `-` may also be used to select the
standard streams explicitly.

If the *STAGE* option is omitted, the compiler currently will encounter an
error because its stages are not yet fully implemented.

	* src/cli.cpp: Implement the new CLI.
	* src/cli.hpp: Updated function interface.
	(real_main): Also accept a `stdin` parameter.
	* src/lexer/lexer.hpp: Remove `noexcept` specifier from
	`default`ed move operations as they would make the program
	ill-formed.
	* src/minijava.cpp: Updated call to `real_main`.
	* tests/cli.cpp: Updated the old tests and added many new.