Make the `Build` script build-type aware.
The default build-type is 'Release' because that's what we want on the
SpeedCenter.  If you want to use the script for your local development, I
recommend that you export the environment variable `CMAKE_BUILD_TYPE=Debug` in
the shell you're normally building in or use the `--build-type` option every
time you run the script.

	* build: Use `CMAKE_BUILD_TYPE` and improved `--help`.