Lexer cleanup 6b of N.
This patch defines the character classification functions independently of
`<ctype.h>` so we are not depending on its idea of what a character is.
Hopefully, this will also have a beneficial effect on performance as our
functions are much simpler.  Speaking of which, however, we should benchmark
whether the `static` initializers impose any overhead and if so, either get
clever or move them into a separate translation unit where they can be global
variables.

	* src/lexer/character.hpp: Remove inline definitions.
	* src/lexer/character.tpp: New.