Allow setting a higher limit for the maximum stack size in the compiler.
This patch teaches the compiler to recognize the environment variable
`MINIJAVA_STACK_LIMIT` that can be set to the string `default`, `none` or a
positive integer to set the limit for the maximum stack size to the system
default, the system maximum or the specified value in bytes, respectively.

The adjustment only works on POSIX.  On other systems, it fails gracefully.
The `run` script sets the limit to `none`.

	* CMakeLists.txt: Check for `<sys/resource.h>`.
	* config.h.in: Add configuration variable.
	(MINIJAVA_HAVE_RLIMIT): New.
	* run: Set `MINIJAVA_STACK_LIMIT` to `NONE`.
	* src/cli.cpp: Adjust stack limit.
	* src/global.hpp: Added environment variable.
	(MINIJAVA_ENVVAR_STACK_LIMIT): New.
	* src/system/rlimit_stack_generic.tpp: New.
	* src/system/rlimit_stack_posix.tpp: New.
	* src/system/system.cpp: New.
	* src/system/system.hpp: New.
	* tests/system/system.cpp: New.