Syntax generator cleanup.
The generator now supports a basic command-line interface that allows selecting
a maximum recursion depth.  The parameters have been tweaked to make the
generated programs look remotely like Java.  The serialization is now a proper
component as we tend to need it over and over again.  It's still not very
polished, though.

	* CMakeLists.txt: Add new component.
	* src/parser/serializer.cpp: New.
	* src/parser/serializer.hpp: New.
	* src/parser/serializer.tpp: New.
	* tests/CMakeLists.txt
	* tests/generator.cpp: The new generator.
	* tests/parser/serializer.cpp: New.
	* tests/parser/syntax_gen.cpp: Was the old generator.