Micro-benchmarks for `token_type_set` operations.
These were non-trivially hard to write as the operation to benchmark is only
two or three instructions.  Therefore, even minor changes to the way, say, the
iteration is written in the benchmark have a large effect on the result.  In
particular, the concrete nanoseconds per operation that are reported should be
considered arbitrary numbers.  The only meaningful comparison is over time
after changes to the `token_type_set` are made.  it is not meaningful to
compare the numbers for, say, insertion and lookup because those use different
loops internally.

	* benchmarks/micro/CMakeLists.txt: Add new benchmarks.
	* benchmarks/micro/Manifest.json: Add new benchmarks.
	* benchmarks/micro/tts-combine.cpp: New.
	* benchmarks/micro/tts-lookup.cpp: New.
	* benchmarks/micro/tts-modify.cpp: New.