Pretty-Printer cleanup.
Use RAII when possible and replace raw loops by algorithms.  There are probably
more places where RAII could be applied but I didn't feel comfortable with
doing so.  A more dense test suite would be needed for this.  Even in the
places where I've applied it, it changed the semantics of the code because most
flags never actually got reset.

	* src/parser/pretty_printer.cpp: Cleanup.