Add integration tests.
This patch adds a new test category for integration tests.  Tests that test the
compiler as a whole.  In order to integrate them, the `tests` directory was
structured to contain now three sub-directories.  The `unit` sub-directory
contains what was formerly directly below `tests` except for the fuzzers, which
now have their own sub-directory `fuzz` and are also run as part of the test
suite.  The integration tests are in the `comp` sub-directory.  In order to add
new integration tests, simply drop the input files in the respective directory.
The test driver will pick them up.

The integration tests for semantic errors pass the `--escape` flag to the test
driver.  This willl make sure that only the piece of code that is commented-out
using this syntax

    void foo() {
        /*** return 5; ***/
    }

triggers the diagnostic and not other random errors that might have sneaked
into the test file.

In order to be able to run the sementic tests, I quickly added the `--check`
option into `cli.cpp`.  This will cause a severe merge conflict and should be
considered throw-away code.

The "unit" tests in the `semantic` sub-module that actually exercised the whole
compiler were converted into integration tests.  There are a few cases
remaining for which more interesting stuff than wether or not they are accepted
is tested and those still need to be refactored to bypass the lexer and parser.

	* README.md: Updated for new directories.
	* src/core/cli.cpp: Quickly add --check option.
	* tests/CMakeLists.txt: New.
	* tests/comp/CMakeLists.txt: New.
	* tests/comp/driver.py: New.
	* tests/comp/lex/invalid/*.mj: : New.
	* tests/comp/lex/valid/*.mj: New.
	* tests/comp/sem/invalid/*.mj: New.
	* tests/comp/syn/invalid/*.mj: New.
	* tests/comp/syn/valid/*.mj: New.
	* tests/fuzz/CMakeLists.txt: Run fuzzers as part of test.
	* tests/unit/semantic/ref_type_analysis.cpp: Move integration tests out.
	* tests/unit/semantic/type_system.cpp: Move integration tests out.
	* tests/unit/CMakeLists.txt: Move stuff around.