Fully annotate the built-in AST.
If the built-in AST is not fully annotated, the overall annotation result is
inconsistent and we break our contract, risking clients to follow the
annotation pointers as documented only to trap into a broken link.  This patch
effectively replaces the call to `perform_shallow_name_type_analysis` with a
call to `perform_full_name_type_analysis` which was renamed to just
`perform_name_type_analysis`.

	* src/core/semantic/name_type_analysis.cpp: Updated signatures.
	* src/core/semantic/name_type_analysis.hpp:
	(perform_full_name_type_analysis): Take a flag whether `main` is
	expected.  Renamed.  Expanded DocString.
	(perform_shallow_name_type_analysis): Removed from the header as
	any possible use from outside seems to be wrong.
	* src/core/semantic/semantic.tpp: Always use full analysis.
	* tests/unit/semantic/name_type_analysis.cpp: Update legacy callers.
	* tests/unit/semantic/semantic.cpp: Uncomment test case that
	exposes the bug.