Treat `-2147483648` and `-(2147483648)` as required.
This patch adds a "negative" attribute to AST nodes of type `integer_constant`.
If the integer literal is *immediately* preceeded by a unary minus operator,
the operator is swallowed and the "negative" flag is set.  In all other cases,
a node for a unary minus is created as it used to.  Since this means that the
AST no longer contains nodes with a value that is outside the range of a 32 bit
integer, we can now store the values directly in the target type which is much
nicer.

	* src/core/parser/ast.hpp:
	(integer_constant): Add "negative" flag.
	* src/core/parser/parser.tpp: Use new flag and do some cleanup.
	* src/core/parser/pretty_printer.cpp: Print parens around integer literals.
	* src/core/semantic/constant.cpp: Take advantage of the changes.
	* src/core/semantic/constant.hpp: Remove unneeded typedefs.
	* src/support/testaux/ast_test_factory.cpp: More convenience functions.
	* src/support/testaux/ast_test_factory.hpp: Likewise.
	* tests/unit/cli.cpp: Remove test for `--print-ast` that is no longer possible.
	* tests/unit/parser/parser.cpp: Test new nodes.
	* tests/unit/parser/pretty_printer.cpp: Update for new format.
	* tests/unit/semantic/constant.cpp: Add tests.