Implement xml_printer and AST comparison based upon it.
This frees us from having to worry about breaking[*] unit tests every time we
touch the pretty-printer or any AST internals.  The XML generation is nowhere
near efficient but fully deterministic even in the face of randomly ordered AST
nodes (for which order doesn't matter).  It achieves this by first serializing
each child node and then sorting the serialization outputs in lexicographic
order.  The XML is not intended to be used for anything else except comparing
ASTs but it probably could be, should this be considered valuable in the
future.  At least, it is formatted in a human-readable way.

[*] There are two ways to break unit tests.  The benign one is to cause false
    positives.  This is easily handled.  The more subtle one is to silently
    make the condition the test asserts on a constant.  (Which might cause
    false negatives later on.)  This has happened at least once in the past
    where the pretty printer would produce the same output for the correct and
    the wrong AST.

	* src/core/cli.cpp: Updated caller.
	* src/core/parser/ast.hpp: Imporove DocString.
	* src/core/parser/ast_misc.cpp: Implement stuff.
	* src/core/parser/ast_misc.hpp: Rename `to_string` to `to_text`.
	* src/support/testaux/ast_test_factory.cpp: More convenience.
	* src/support/testaux/ast_test_factory.hpp: Likewise.
	* tests/fuzz/print-ast.cpp: Updated caller.
	* tests/unit/parser/ast_misc.cpp: Add tests.
	* tests/unit/parser/pretty_printer.cpp: Refactor unit tests.