Rudimentary version of runtime support library and MiniJava to C backend.
This patch adds the runtime support library `libruntime` that can be linked in
to provide the C functions `mj_runtime_allocate` and `mj_runtime_println`.
Their implementation is currently rather crude and not optimized at all.

By far the bigger contribution from this patch is a C backend for the MiniJava
compiler that can be used to test the runtime support library and actually the
whole compiler.  It generates C99 code that can be compiled with GCC and linked
against `libruntime`.  Ideally, compiling MiniJava with `javac` (where it is
compatible), compiling it to C and compiling it native should all produce
programs that behave the same.

Assuming you've built in the `stage` subdirectory, then you can use the C
backend like so.

    $ ./stage/src/mj2c < example.mj > example.c
    $ gcc -c example.c
    $ gcc -o example example.o ./stage/src/runtime/libruntime.a

It might be desirable to compile with `-g` and `-Og` and enable
`-fsanitize=address` or `-fsanitize=undefined`.  Note that AddressSanitizer
will always complain about memory leaks in `libruntime`, though, as it never
frees the memory again.

	* src/CMakeLists.txt: Add new executable `mj2c`.
	* src/mj2c.cpp: New.
	* src/runtime/CMakeLists.txt: New.
	* src/runtime/allocate.c: New.
	* src/runtime/println.c: New.