Fix name-clash pointed out in review.
The code used to use a '$' rather than an '_' for joining names but this was
changed last-minute before I submitted the previous patch to be more
standards-compliant.  This patch now adds the length of identifiers into the
mangled name as suggested by Tobias.  There are still issues with the name
mangling, though.  For example, we formally invoke undefined behavior by
potentially using names with consecutive underscores.  I think that the only
really safe solution would be to simply map identifiers to random strings but
that would produce very unreadable code and the main purpose of this backend is
for debugging.

	* src/mj2c.cpp: Improve name mangling and don't depend on
	<stdlib.h> just for NULL.