Use a `constexpr` perfect Pearson hash function for keyword classification.
The Pearson table is generated by a Python script which is called by CMake.
The hash function will probably not work with non-ASCII character sets (as far
as the characters that may occur in a keyword are concerned) but this is
statically asserted on and the unit tests check the proper functioning for
every keyword.

	* extras/gen/keyword-hash.py: New (generator script).
	* src/core/CMakeLists.txt: Call the generator.
	* src/core/lexer/keyword.cpp: Remove old code.
	* src/core/lexer/keyword.hpp: `#include` new `*.tpp` files.
	* src/core/lexer/keyword.tpp: New.
	* tests/lexer/keyword.cpp: Additional `constexpr` tests.