Expose `is_number` and `is_word` functions in `lexer/character.hpp`.
While these functions are never actually needed in the lexer, there is a
frequent need for them in functions that are supposed to be only called with
valid words or numbers and wish to assert these preconditions.  Defining and
unit-testing these helper functions in the one right place is better than
redundantly defining poor alternatives in several places or resorting to regex
and other questionable solutions on an ad hoc basis.

	* src/core/lexer/character.cpp: Add new functions.
	* src/core/lexer/character.hpp: Likewise.
	* tests/lexer/character.cpp: Likewise.