Make the RAII handling of libfirm a bit more sane.
I cannot resist the temptation to put an analogue of the following comment
here, just in case you think it is still not sane enough...

    http://stackoverflow.com/a/482129

Any function that operates on a Firm IRP now asserts that libfirm is still
initialized.  The global libfirm state wrapper can be used to obtain IRPs which
are counted such that we can assert none get leaked.  Since libfirm doesn't
allow setting the global IRP to NULL, the implicitly created IRP is used for
that purpose instead.  You can use it but our code never does.  If anybody
messes with libfirm's global state by means other than through our interface,
we should be able to catch it and reliably throw an exception.

	* src/core/cli.cpp:
	* src/core/irg/global_firm_state.cpp:
	* src/core/irg/global_firm_state.hpp:
	* src/core/irg/irg.cpp:
	* src/core/irg/irg.hpp:
	* src/core/irg/lowering.cpp:
	* src/core/irg/lowering.hpp:
	* tests/comp/dump-ir/field_access.mj:
	* tests/unit/irg/global_firm_state.cpp:
	* tests/unit/irg/irg.cpp:
	* tests/unit/irg/mangle.cpp:
	* tests/unit/irg/type_builder.cpp: