Refactoring of integration test driver.
Instead of command-line flags -- which turned out to be a cumbersome technique
for this -- the test driver is now mostly controlled by a simple pre-processor
language (which is modeled after the C pre-processor) directly in the source
files that are the inputs for the tests.  If I have time, I might write a small
HOWTO later but the existing tests should be pretty self-explanatory.  The
pre-processor language has simple conditionals and pragmas to control the test
driver.  Instead of one directory with tests per compiler stage, we now have
only two directories any more: `analyze` and `execute`.  The former contains
programs that should never be executed.  The latter contains only semantically
correct programs with no undefined behavior at run-time.  While it would also
be possible to `#ifdef` around this and only use a single directory of tests,
the separation seems to make enough sense to justify it.  Other than that, all
tests are always run trough all stages and pragmas select at what stage
negative tests should fail.  (So to make sure that a test for a semantic error
doesn't actually fail due to an accidental syntax error in the same file.)
Right now, we don't actually compile the `analyze` tests because the compiler
is not yet ready for this but once it is, simply add one line to the
`CMakeLists.txt` and get dozens of tests for free.