Refactor handling of runtime library builtins.
It turns out that we don't need to duplicate (or even triplicate) all the
knowledge about the runtime library builtins.  The only place that knows about
them now is the semantic analysis (and the C implementation of the runtime
library, of course).  The Firm IRG construction can be agnostic of it.

This is step one in preparing for adding more runtime library features.

	* src/core/irg/method_builder.cpp: Use generic code to handle runtime library builtins.
	* src/core/irg/method_builder.hpp: Likewise.
	* src/core/irg/type_builder.cpp: Likewise.
	* src/core/irg/type_builder.hpp: Likewise.
	* src/core/runtime/runtime.cpp: Rename `allocate` to `new` to avoid name clashes.