Move runtime library to its own source file (again).
Maintaining C code embedded as string literals in C++ code turned out to be not
very workable.  Especially with editors that get confused by the syntax.  This
change also allows us to compile the runtime library in isolation for testing
it.

The `extras/gen/blobdump.py` script that converts the C source file to string
literals is generic and can be used in the future to convert other files as
well.

	* extras/gen/blobdump.py: New.
	* src/CMakeLists.txt: Add sub-directory `runtime`.
	* src/core/CMakeLists.txt: `add_custom_command` for `runtime.tpp`.
	* src/core/runtime/runtime.cpp: Remove inline code.
	* src/runtime/CMakeLists.txt: New.
	* src/runtime/mj_runtime.c: New.