Add `int System.id(int x)` runtime library builtin.
This function simply returns its argument unchanged.  It is intended to be used
as a barrier to stop the compiler from propagating constants in places where
doing so would defeat the purpose of a test.

If you ever implement a compiler "optimization" that specifically inlines this
builtin, I'll tell Santa Claus what you did.

	* src/core/semantic/semantic.tpp: Add to builtin AST.
	* src/runtime/mj_runtime.c: Add function.
	(mj_runtime_id): New.
	* tests/comp/execute/system_id.mj: New.