Make support for line comments an optional extension as per supervisor request.
Support for line comments can be activated through the `MINIJAVA_LINE_COMMENTS`
configuration option but is disabled by default.  If nobody is willing to set
up yet another test build that also checks our compiler with this extension
enabled, it might be the smartest thing to simply remove it altogether.

	* CMakeLists.txt: Add MINIJAVA_LINE_COMMENTS option.
	* README.md: Document new option.
	* config.h.in: Add MINIJAVA_LINE_COMMENTS macro.
	* src/core/lexer/lexer.tpp: Honor MINIJAVA_LINE_COMMENTS macro.
	* tests/unit/lexer/lexer.cpp: Likewise.